+++ run_ocp_centroidal.py
@@ -1,9 +1,10 @@
 import time
 import numpy as np
 import pinocchio as pin
+import casadi
 
 from helpers import *
-from ocp_rnea import OptimalControlRNEA
+from ocp_centroidal import OCP_Centroidal
 
 # Problem parameters
 # robot = B2(reference_pose="standing")
@@ -20,37 +21,12 @@ arm_vel_des = np.array([0.1, 0, 0])
 # Tracking goal: linear and angular momentum
 com_goal = np.array([0.1, 0, 0, 0, 0, 0])
 
-# MPC
-mpc_loops = 50
-
-# Compiled solver
+# Solver
+solver = "osqp"
 compile_solver = False
 load_compiled_solver = None
 
 debug = True  # print info
-
-
-def mpc_loop(ocp, robot_instance, q0, N):
-    x_init = np.concatenate((q0, np.zeros(robot_instance.model.nv)))
-    solve_times = []
-
-    # Initialize solver
-    ocp.init_solver(solver="osqp", compile_solver=compile_solver)
-
-    for k in range(N):
-        ocp.update_initial_state(x_init)
-        ocp.update_contact_schedule(shift_idx=k)
-        ocp.warm_start()
-        ocp.solve(retract_all=False)
-        solve_times.append(ocp.solve_time)
-
-        x_init = ocp.dyn.state_integrate()(x_init, ocp.DX_prev[1])
-        robot_instance.display(ocp.qs[-1])  # Display last q
-
-    print("Avg solve time (ms): ", np.average(solve_times) * 1000)
-    print("Std solve time (ms): ", np.std(solve_times) * 1000)
-
-    return ocp
 
 
 def main():
@@ -68,36 +44,65 @@ def main():
 
     pin.computeAllTerms(model, data, q0, np.zeros(model.nv))
 
-    robot_instance.initViewer()
-    robot_instance.loadViewerModel("pinocchio")
-    robot_instance.display(q0)
-
     # Setup OCP
-    ocp = OptimalControlRNEA(
+    ocp = OCP_Centroidal(
         robot=robot,
         nodes=ocp_nodes,
     )
     ocp.set_com_goal(com_goal)
     ocp.set_arm_task(arm_f_des, arm_vel_des)
-    ocp = mpc_loop(ocp, robot_instance, q0, mpc_loops)
+
+    x_init = np.concatenate((np.zeros(6), q0))
+    gait_idx = 0
+
+    if load_compiled_solver:
+        compiled_solver = casadi.external("compiled_solver", "codegen/lib/" + load_compiled_solver)
+        contact_schedule = ocp.gait_sequence.shift_contact_schedule(gait_idx)[:,:ocp_nodes]
+        start_time = time.time()
+        DX_sol, U_sol = compiled_solver(x_init, contact_schedule, gait_idx)
+        end_time = time.time()
+        print("DX_sol: ", DX_sol)
+        print("U_sol: ", U_sol)
+        print("Solve time (ms): ", (end_time - start_time) * 1000)
+        return
+
+    else:
+        ocp.update_initial_state(x_init)
+        ocp.update_contact_schedule(shift_idx=gait_idx)
+        ocp.init_solver(solver=solver, compile_solver=compile_solver)
+        ocp.solve(retract_all=True)
+        print("Solve time (ms):", ocp.solve_time * 1000)
 
     if debug:
-        for k in range(len(ocp.qs)):
+        for k in range(ocp_nodes):
             q = ocp.qs[k]
-            v = ocp.vs[k]
-            tau = ocp.taus[k]
-            forces = ocp.fs[k]
-            print("q: ", q.T)
-            print("v: ", v.T)
-            print("tau: ", tau.T)
+            h = ocp.hs[k]
+            u = ocp.us[k]
 
-            if k < len(ocp.vs) - 1:
-                v_next = ocp.vs[k + 1]
+            forces = u[:robot.nf]
+            v_j = u[robot.nf:]
+            v_b = ocp.dyn.get_base_velocity()(h, q, v_j)
+            v_b = np.array(v_b).flatten()
+            v = np.concatenate((v_b, v_j))
+            pin.computeAllTerms(model, data, q, v)
+
+            print("k: ", k)
+            print("h: ", h.T)
+            print("q: ", q.T)
+            print("h true: ", data.hg / data.mass[0])
+
+            # RNEA
+            if k < ocp_nodes - 1:
+                h_next = ocp.hs[k + 1]
+                q_next = ocp.qs[k + 1]
+                v_j_next = ocp.us[k + 1][robot.nf:]
+                v_b_next = ocp.dyn.get_base_velocity()(h_next, q_next, v_j_next)
+                v_b_next = np.array(v_b_next).flatten()
+                v_next = np.concatenate((v_b_next, v_j_next))
             else:
                 break
             a = (v_next - v) / dt
 
-            # RNEA
             pin.framesForwardKinematics(model, data, q)
             f_ext = [pin.Force(np.zeros(6)) for _ in range(model.njoints)]
             for idx, frame_id in enumerate(robot.ee_ids):
@@ -113,16 +118,16 @@ def main():
                 f_ext[joint_id] = pin.Force(f)
 
             tau_rnea = pin.rnea(model, data, q, v, a, f_ext)
-
-            tau_total = np.concatenate((np.zeros(6), tau))
-            print("tau gap: ", tau_total - tau_rnea)
+            print("tau rnea: ", tau_rnea)
 
     # Visualize
+    robot_instance.initViewer()
+    robot_instance.loadViewerModel("pinocchio")
     for _ in range(50):
-        for k in range(len(ocp.qs)):
-            q = ocp.qs[k]
-            robot_instance.display(q)
+        for q in ocp.qs:
+            robot_instance.display(q)        
             time.sleep(dt)
+
 
 if __name__ == "__main__":
     main()
