+++ run_ocp_rnea.py
@@ -1,10 +1,9 @@
 import time
 import numpy as np
 import pinocchio as pin
-import casadi
 
 from helpers import *
-from optimal_control_problem import OptimalControlProblem
+from ocp_rnea import OCP_RNEA
 
 # Problem parameters
 # robot = B2(reference_pose="standing")
@@ -21,7 +20,7 @@ arm_vel_des = np.array([0.1, 0, 0])
 # Tracking goal: linear and angular momentum
 com_goal = np.array([0.1, 0, 0, 0, 0, 0])
 
-# Solver
+# Compiled solver
 solver = "osqp"
 compile_solver = False
 load_compiled_solver = None
@@ -45,64 +44,38 @@ def main():
     pin.computeAllTerms(model, data, q0, np.zeros(model.nv))
 
     # Setup OCP
-    ocp = OptimalControlProblem(
+    ocp = OCP_RNEA(
         robot=robot,
         nodes=ocp_nodes,
     )
     ocp.set_com_goal(com_goal)
     ocp.set_arm_task(arm_f_des, arm_vel_des)
 
-    x_init = np.concatenate((np.zeros(6), q0))
+    x_init = np.concatenate((q0, np.zeros(model.nv)))
     gait_idx = 0
 
-    if load_compiled_solver:
-        compiled_solver = casadi.external("compiled_solver", "codegen/lib/" + load_compiled_solver)
-        contact_schedule = ocp.gait_sequence.shift_contact_schedule(gait_idx)[:,:ocp_nodes]
-        start_time = time.time()
-        DX_sol, U_sol = compiled_solver(x_init, contact_schedule, gait_idx)
-        end_time = time.time()
-        print("DX_sol: ", DX_sol)
-        print("U_sol: ", U_sol)
-        print("Solve time (ms): ", (end_time - start_time) * 1000)
-        return
-
-    else:
-        ocp.update_initial_state(x_init)
-        ocp.update_contact_schedule(shift_idx=gait_idx)
-        ocp.init_solver(solver=solver, compile_solver=compile_solver)
-        ocp.solve(retract_all=True)
-        print("Solve time (ms):", ocp.solve_time * 1000)
+    ocp.update_initial_state(x_init)
+    ocp.update_contact_schedule(shift_idx=gait_idx)
+    ocp.init_solver(solver=solver, compile_solver=compile_solver)
+    ocp.solve(retract_all=True)
 
     if debug:
         for k in range(ocp_nodes):
             q = ocp.qs[k]
-            h = ocp.hs[k]
-            u = ocp.us[k]
-
-            forces = u[:robot.nf]
-            v_j = u[robot.nf:]
-            v_b = ocp.dyn.get_base_velocity()(h, q, v_j)
-            v_b = np.array(v_b).flatten()
-            v = np.concatenate((v_b, v_j))
-            pin.computeAllTerms(model, data, q, v)
-
-            print("k: ", k)
-            print("h: ", h.T)
+            v = ocp.vs[k]
+            tau = ocp.taus[k]
+            forces = ocp.fs[k]
             print("q: ", q.T)
-            print("h true: ", data.hg / data.mass[0])
+            print("v: ", v.T)
+            print("tau: ", tau.T)
 
-            # RNEA
             if k < ocp_nodes - 1:
-                h_next = ocp.hs[k + 1]
-                q_next = ocp.qs[k + 1]
-                v_j_next = ocp.us[k + 1][robot.nf:]
-                v_b_next = ocp.dyn.get_base_velocity()(h_next, q_next, v_j_next)
-                v_b_next = np.array(v_b_next).flatten()
-                v_next = np.concatenate((v_b_next, v_j_next))
+                v_next = ocp.vs[k + 1]
             else:
                 break
             a = (v_next - v) / dt
 
+            # RNEA
             pin.framesForwardKinematics(model, data, q)
             f_ext = [pin.Force(np.zeros(6)) for _ in range(model.njoints)]
             for idx, frame_id in enumerate(robot.ee_ids):
@@ -120,14 +93,19 @@ def main():
             tau_rnea = pin.rnea(model, data, q, v, a, f_ext)
             print("tau rnea: ", tau_rnea)
 
+            tau_total = np.concatenate((np.zeros(6), tau))
+            print("tau gap: ", tau_total - tau_rnea)
+
+    print("Solve time (ms):", ocp.solve_time * 1000)
+
     # Visualize
     robot_instance.initViewer()
     robot_instance.loadViewerModel("pinocchio")
-    for _ in range(50):
-        for q in ocp.qs:
-            robot_instance.display(q)        
+    for _ in range(100):
+        for k in range(ocp_nodes):
+            q = ocp.qs[k]
+            robot_instance.display(q)
             time.sleep(dt)
-
 
 if __name__ == "__main__":
     main()
